requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
chardet==3.0.4
colorama==0.4.4
DateTimeRange==1.0.0
Django==3.1.3
django-phonenumber-field==5.0.0
isort==5.6.4
lazy-object-proxy==1.4.3
mbstrdecoder==1.0.0
mccabe==0.6.1
phonenumbers==8.12.13
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typepy==1.1.1
wrapt==1.12.1
